\docType{package}
\name{datamerge-package}
\alias{datamerge}
\alias{datamerge-package}
\title{Merge partially overlapping and inconsistent data}
\description{
  Merge partially overlapping and inconsistent data
}
\details{
  \tabular{ll}{ Package: \tab datamerge\cr Type: \tab
  Package\cr Version: \tab 1.0-1\cr Date: \tab
  2012-07-23\cr License: \tab GPL\cr LazyLoad: \tab yes\cr
  Depends: \tab xtermStyle\cr }
}
\author{
  Christofer \enc{Bäcklin}{Backlin}
  \url{http://christofer.backlin.se/R/}
}
\seealso{
  \code{\link{version.merge}}, \code{\link{clean.factors}}
}
\keyword{data}
\keyword{merge}
\keyword{versioned}

